+++ Tests/RationalNumberTests.swift
@@ -49,22 +49,3 @@ class RationalNumberTests: XCTestCase, TypeTest {
 		doubles(values: [0.125, 0.00125, 2.0.sqrt, 3.0.sqrt, 3e32.sqrt, nextafter(0.0, DBL_MAX) ])
 	}
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
